# admin-ui/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "arctern",
    "description": "ZFS replication daemon HTTP API",
    "license": {
      "name": ""
    },
    "version": "0.3.0"
  },
  "paths": {
...
    "/api/v1/peers/{peer}/events": {
      "get": {
        "tags": [
          "peers"
        ],
        "summary": "`GET /api/v1/peers/{peer}/events` — proxied SSE. Sends\nSubscribeEvents to the peer's control channel and yields each\npushed Event frame as an SSE frame.",
        "operationId": "stream_peer_events",
        "parameters": [
          {
            "name": "peer",
            "in": "path",
            "description": "Peer name from [[peers]]",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "SSE stream of LogEvent JSON frames from the peer"
          },
          "404": {
            "description": "No such peer",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "503": {
            "description": "Peer not currently connected",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/peers/{peer}/jobs/stream": {
      "get": {
        "tags": [
          "peers"
        ],
        "summary": "Peer-scoped counterpart of `/api/v1/jobs/stream`. The receiver's generic\ncontrol RPC returns bounded JSON replies, so this bridge samples it over\nthe already-open SSH control channel and exposes one long-lived SSE\nresponse to the browser. Unchanged snapshots are not emitted.",
        "operationId": "stream_peer_jobs",
        "parameters": [
          {
            "name": "peer",
            "in": "path",
            "description": "Peer name from [[peers]]",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "SSE stream of job-status snapshots from the peer"
          },
          "404": {
            "description": "No such peer",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "503": {
            "description": "Peer not currently connected",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/pools": {
      "get": {
        "tags": [
          "pools"
        ],
        "operationId": "list_pools",
        "responses": {
          "200": {
            "description": "All imported pools with health + capacity",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PoolSummary"
                  }
                }
              }
            }
          },
          "500": {
            "description": "zpool returned an error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/pools/{name}": {
      "get": {
        "tags": [
          "pools"
        ],
        "operationId": "get_pool",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "description": "Pool name",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Full status with vdev tree + scrub progress",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PoolStatus"
                }
              }
            }
          },
          "404": {
            "description": "No imported pool with that name",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "500": {
            "description": "zpool returned an error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/pools/{name}/scrub": {
      "post": {
        "tags": [
          "pools"
        ],
        "operationId": "pool_scrub",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "description": "Pool name",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScrubRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "zpool scrub accepted the action"
          },
          "400": {
            "description": "Unknown action",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "500": {
            "description": "zpool returned an error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ApiErrorBody": {
        "type": "object",
        "description": "Body shape for `4xx`/`5xx` responses from the daemon. `error` is a\nshort machine-readable category (`spawn`, `dataset_not_found`, …);\n`message` is a human-readable description.",
        "required": [
          "error",
          "message"
        ],
        "properties": {
          "error": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        }
      },
...
      "PoolStatus": {
        "type": "object",
        "description": "`GET /api/v1/pools/{name}` — full status: scrub + recursive vdev tree.",
        "required": [
          "name",
          "state",
          "error_count",
          "pool_guid",
          "txg",
          "vdevs"
        ],
        "properties": {
          "name": {
            "type": "string"
          },
          "state": {
            "type": "string"
          },
          "error_count": {
            "type": "string"
          },
          "pool_guid": {
            "type": "string"
          },
          "txg": {
            "type": "string"
          },
          "scan": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/ScanSummary"
              }
            ]
          },
          "vdevs": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/VdevNode"
            }
          }
        }
      },
      "PoolSummary": {
        "type": "object",
        "description": "One pool's slot in `GET /api/v1/pools`. Numeric fields are\n`zpool`-formatted strings (e.g. `\"608G\"`, `\"1.48T\"`) rather than\nraw bytes because that's what `zpool` emits and round-tripping\nthrough bytes risks rounding mismatches in the UI.",
        "required": [
          "name",
          "state",
          "error_count",
          "alloc_space",
          "total_space"
        ],
        "properties": {
          "name": {
            "type": "string"
          },
          "state": {
            "type": "string",
            "description": "`\"ONLINE\"`, `\"DEGRADED\"`, `\"FAULTED\"`, …"
          },
          "error_count": {
            "type": "string",
            "description": "Aggregate error count across all vdevs."
          },
          "alloc_space": {
            "type": "string"
          },
          "total_space": {
            "type": "string"
          },
          "scan": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/ScanSummary",
                "description": "Most recent scrub/resilver status if zpool reports one."
              }
            ]
          }
        }
      },
...
      "ScanSummary": {
        "type": "object",
        "required": [
          "function",
          "state"
        ],
        "properties": {
          "function": {
            "type": "string",
            "description": "`\"SCRUB\"`, `\"RESILVER\"`, `\"NONE\"`."
          },
          "state": {
            "type": "string",
            "description": "`\"SCANNING\"`, `\"FINISHED\"`, `\"CANCELED\"`."
          },
          "start_time": {
            "type": [
              "string",
              "null"
            ]
          },
          "end_time": {
            "type": [
              "string",
              "null"
            ]
          },
          "to_examine": {
            "type": [
              "string",
              "null"
            ]
          },
          "examined": {
            "type": [
              "string",
              "null"
            ]
          },
          "errors": {
            "type": [
              "string",
              "null"
            ]
          },
          "pass_start": {
            "type": [
              "string",
              "null"
            ]
          },
          "scrub_pause": {
            "type": [
              "string",
              "null"
            ]
          },
          "issued": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
      "ScrubAction": {
        "type": "string",
        "description": "`zpool scrub` verbs.",
        "enum": [
          "start",
          "pause",
          "resume",
          "stop"
        ]
      },
      "ScrubRequest": {
        "type": "object",
        "description": "Body of `POST /api/v1/pools/{name}/scrub`.",
        "required": [
          "action"
        ],
        "properties": {
          "action": {
            "$ref": "#/components/schemas/ScrubAction"
          }
        }
      },
...
      "VdevNode": {
        "type": "object",
        "description": "Recursive vdev tree as a flat list of trees. Wire-friendlier than\nzfskit's map<name, VdevStatus> for UIs that want to render in\ndeclared order.\n\n`children` carries the `#[schema(no_recursion)]` attribute so\nutoipa's auto-collector stops at the cycle and emits a `$ref` to\n`VdevNode` itself instead of inlining the type — without this\n`ApiDoc::openapi()` infinite-recurses and overflows the stack at\nstartup. See utoipa docs on recursive schemas.",
        "required": [
          "name",
          "vdev_type",
          "state",
          "alloc_space",
          "total_space",
          "read_errors",
          "write_errors",
          "checksum_errors",
          "children"
        ],
        "properties": {
          "name": {
            "type": "string"
          },
          "vdev_type": {
            "type": "string"
          },
          "state": {
            "type": "string"
          },
          "alloc_space": {
            "type": "string"
          },
          "total_space": {
            "type": "string"
          },
          "read_errors": {
            "type": "string"
          },
          "write_errors": {
            "type": "string"
          },
          "checksum_errors": {
            "type": "string"
          },
          "path": {
            "type": [
              "string",
              "null"
            ]
          },
          "children": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/VdevNode"
            }
          }
        }
      }
    }
  }
}